Bug fixes, Prettier formatting, string building
Transformers icons stopped showing, so now it uses the raw data as a
more reliable source. The card background pseudo element went outside
the card, which is now kept inside with display:inline-block and slight
but precise adjustments were needed.

Moved over to working in VS Code with live compiling and live server,
which makes things easier. That bringed on the html source of the css
testing grounds and some Prettier auto formatting

Improve upon how the result strings in many functions are built
together and add some error reporting